+++ lab8/user/libs/syscall.h
@@ -3,7 +3,7 @@
 
 int sys_exit(int64_t error_code);
 int sys_fork(void);
-int sys_wait(int64_t pid, int64_t *store);
+int sys_wait(int64_t pid, int *store);
 int sys_exec(const char *name, int64_t argc, const char **argv);
 int sys_yield(void);
 int sys_kill(int64_t pid);
